+++ pkg/providers/mtlstruststore/testdata/TestDataCASetAssociations/id.tf
@@ -1,0 +1,7 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_mtlstruststore_ca_set_associations" "test" {
+  id = "123"
+}